added user

+++ src/repository/blogRepo.js
@@ -1,9 +1,11 @@
+const ErrorResponse = require("../utils/errorResponse");
 const BlogModel = require("../models/Blog");
 
-exports.create = async ({ title, content }) => {
+exports.create = async ({ title, content, author }) => {
 	const blog = new BlogModel({
 		title,
 		content,
+		author,
 	});
 	return blog.save();
 };
@@ -18,13 +20,20 @@ exports.getAllBlogs = async () => {
 	return blogs;
 };
 
-exports.updateBlog = async (id, { title, content }) => {
+exports.updateBlog = async (id, { title, content }, author) => {
 	const blog = await BlogModel.findById(id);
+	if (blog.author !== author) {
+		throw new ErrorResponse("You are not authorized to update this blog", 401);
+	}
 	blog.title = title;
 	blog.content = content;
 	await blog.save();
 };
 
-exports.deleteBlog = async (id) => {
+exports.deleteBlog = async (id, author) => {
+	const blog = await BlogModel.findById(id);
+	if (blog.author !== author) {
+		throw new ErrorResponse("You are not authorized to update this blog", 401);
+	}
 	return await BlogModel.deleteOne({ id });
 };
